+++ src/pages/club/main/ClubRecruitCard/style.tsx
@@ -1,17 +1,15 @@
 import styled from 'styled-components';
-import { clubRecruitCard } from '@/types/clubRecruitCard';
-import { useState } from 'react';
 import { Star as EmptyStarIcon } from '@styled-icons/boxicons-regular/Star';
 import { Star as FilledStarIcon } from '@styled-icons/boxicons-solid/Star';
 
-const MainContainer = styled.div`
+export const MainContainer = styled.div`
 	width: 62.5rem;
 	display: flex;
 	flex-direction: column;
 	align-items: center;
 	justify-content: center;
 `;
-const ClubRecuitContainer = styled.div`
+export const ClubRecuitContainer = styled.div`
 	width: 62.5rem;
 	height: 6.25rem;
 	flex-shrink: 0;
@@ -24,19 +22,19 @@ const ClubRecuitContainer = styled.div`
 	background: #fff;
 `;
 
-const ClubNameDateContainer = styled.div`
+export const ClubNameDateContainer = styled.div`
 	width: 15rem;
 	display: flex;
 	flex-direction: column;
 	align-items: center;
 `;
-const ClubDescriptionContainer = styled.div`
+export const ClubDescriptionContainer = styled.div`
 	width: 36.5rem;
 	display: flex;
 	flex-direction: column;
 	align-items: flex-start;
 `;
-const ClubDescription = styled.p`
+export const ClubDescription = styled.p`
 	color: #585858;
 	text-align: start;
 	font-size: 0.875rem;
@@ -48,21 +46,21 @@ const ClubDescription = styled.p`
 	-webkit-box-orient: vertical;
 	overflow: hidden;
 `;
-const VerticalLine = styled.div`
+export const VerticalLine = styled.div`
 	width: 0.0625rem;
 	height: 4.375rem;
 	margin-right: 1.5rem;
 	background: #d9d9d9;
 `;
 
-const ClubDeadline = styled.p`
+export const ClubDeadline = styled.p`
 	color: #585858;
 	font-size: 0.6875rem;
 	font-weight: 400;
 	line-height: 1.375rem;
 	letter-spacing: 0.0625rem;
 `;
-const ClubSubmitButton = styled.button`
+export const ClubSubmitButton = styled.button`
 	width: 7rem;
 	height: 2.25rem;
 	flex-shrink: 0;
@@ -73,7 +71,7 @@ const ClubSubmitButton = styled.button`
 	color: #fff;
 `;
 
-const ClubDeadlineCounter = styled.div`
+export const ClubDeadlineCounter = styled.div`
 	width: 2.75rem;
 	height: 1.25rem;
 	flex-shrink: 0;
@@ -87,41 +85,42 @@ const ClubDeadlineCounter = styled.div`
 	font-weight: 800;
 	line-height: 150%;
 `;
-const ClubRecruitName = styled.p`
+export const ClubRecruitName = styled.p`
 	color: #000;
 	margin: 0.75rem 0;
 	font-size: 1.125rem;
 	font-weight: 700;
 	letter-spacing: 0.0625rem;
 `;
-const ClubApplicantCount = styled.p`
+export const ClubApplicantCount = styled.p`
 	color: #585858;
 	text-align: center;
 	font-size: 0.6875rem;
 	font-weight: 400;
 	letter-spacing: 0.0625rem;
+	span {
+		color: #3574f2;
+		font-size: 0.6875rem;
+		font-weight: 700;
+		letter-spacing: 0.0625rem;
+	}
 `;
-const ClubApplicantCountSpan = styled.span`
-	color: #3574f2;
-	font-size: 0.6875rem;
-	font-weight: 700;
-	letter-spacing: 0.0625rem;
-`;
-const ToggleButton = styled.button`
+
+export const ToggleButton = styled.button`
 	height: 1.5rem;
 	border: none;
 	border-radius: 0 0 10px 10px;
 	color: #000;
 	background: none;
 `;
-const DeadLineWrapper = styled.div`
+export const DeadLineWrapper = styled.div`
 	width: 100%;
 	display: flex;
   position: relative;
 	align-items: center;
 	justify-content: center;
 `;
-const StarButton = styled.button`
+export const StarButton = styled.button`
 	position: absolute;
 	display: inline-block;
 	left: 0;
@@ -131,13 +130,13 @@ const StarButton = styled.button`
 	border: none;
 	color: #fff;
 `;
-const EmptyStar = styled(EmptyStarIcon)`
+export const EmptyStar = styled(EmptyStarIcon)`
 	width: 1.625rem;
 	height: 1.625rem;
 	color: #9c9c9c;
 	background-color: #fff;
 `;
-const FilledStar = styled(FilledStarIcon)`
+export const FilledStar = styled(FilledStarIcon)`
 	width: 1.625rem;
 	height: 1.625rem;
 	color: #fbbc05;
@@ -146,11 +145,11 @@ const FilledStar = styled(FilledStarIcon)`
 interface StyledPreviewProps {
 	$visible: boolean;
 }
-const activeExist = ({ $visible = true }: StyledPreviewProps) => {
+export const activeExist = ({ $visible = true }: StyledPreviewProps) => {
 	return `height: ${$visible ? '19rem' : '1.5rem'}`;
 };
 
-const PreviewWrapper = styled.div<StyledPreviewProps>`
+export const PreviewWrapper = styled.div<StyledPreviewProps>`
 	display: flex;
 	flex-direction: column;
 	width: 62.5rem;
@@ -161,47 +160,3 @@ const PreviewWrapper = styled.div<StyledPreviewProps>`
 	transition: height 0.3s ease-in-out;
 	background: #e5e5e5;
 `;
-interface Props {
-	recruitInfo: clubRecruitCard;
-}
-export default function ClubRecruitCard({ recruitInfo }: Props) {
-	const [visible, seVisible] = useState(false);
-	const [isStartClicked, setStarClicked] = useState(false);
-
-	const handleClick = () => {
-		setStarClicked(!isStartClicked);
-	};
-	const endDate: Date = new Date(recruitInfo.endDate);
-	const today: Date = new Date();
-	const timeDifference: number = endDate.getTime() - today.getTime();
-	const daysRemaining: number = Math.ceil(timeDifference / (1000 * 60 * 60 * 24));
-
-	
-	return (
-		<MainContainer>
-			<ClubRecuitContainer>
-				<ClubNameDateContainer>
-					<DeadLineWrapper>
-						<StarButton onClick={handleClick}>{!isStartClicked ? <EmptyStar /> : <FilledStar />}</StarButton>
-						<ClubDeadlineCounter>Day-{daysRemaining > 0 ? daysRemaining : 0}</ClubDeadlineCounter>
-					</DeadLineWrapper>
-					<ClubRecruitName>{recruitInfo.title}</ClubRecruitName>
-					<ClubApplicantCount>
-						지원자 : <ClubApplicantCountSpan>{recruitInfo.applicantCount}</ClubApplicantCountSpan>명
-					</ClubApplicantCount>
-				</ClubNameDateContainer>
-				<VerticalLine />
-				<ClubDescriptionContainer>
-					<ClubDescription>{recruitInfo.description}</ClubDescription>
-					<ClubDeadline>
-						{recruitInfo.startDate} ~ {recruitInfo.endDate}
-					</ClubDeadline>
-				</ClubDescriptionContainer>
-				<ClubSubmitButton>지원하기</ClubSubmitButton>
-			</ClubRecuitContainer>
-			<PreviewWrapper $visible={visible}>
-				<ToggleButton onClick={() => seVisible(!visible)}>미리보기</ToggleButton>
-			</PreviewWrapper>
-		</MainContainer>
-	);
-}
